server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "pymongo>=4.16.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
